Updated to SDK-type project and .NET 5.0.
This crashes immediately upon start, looking for:
File name: 'Xwt.Gtk, Version=0.1.0.0, Culture=neutral, PublicKeyToken=null'
As the copy of Xwt.Gtk.dll included with TrueCraft has an informational
version of 0.1.0.0-prerelease, and not 0.1.0.0, I think this is
causing the problem.  Perhaps the newer runtime checks versions
differently or more strictly.  In any case, the Xwt needs to be updated.