Fix: StackOverflowException due to lighting.
Lighting a chunk was causing neighboring chunks to be loaded if they
had previously been generated.  In turn, this would cause these to
be lighted.  If a sufficient number of chunks were previously generated,
this would lead to a StackOverflowException.  Now, the neighbouring
chunks are only lighted if they are already loaded into memory.