Fix: receiving network packets.
When a receive operation completed synchronously, OperationCompleted was called immediately.
This means that network packets were processed out of order, leading to errors.
Since the next Receive operation is no longer started until the current one is processed,
there is no need for a semaphore to ensure only one thread is receiving at a time.