Remove TestCornerCollision.
It is exceptionally unlikely that the exact corner of an Entity's bounding
box will collide with the exact corner of a block.
For this sort of corner collision to happen 4 double values have to have
their 53-bit mantissas be at precisely the right values.
If this were to happen it would be acceptable for the collision to be
treated as a collision with a single face of the block.