Add new unit tests for Ray interections
There is a new interpretation of intersection that has changed some existing
tests.  If the Ray begins on the surface of the box and points outward, it is
no longer considered an intersection.  This avoids entities sticking to walls.
New tests ensure that beginning on the surface of the box and pointing inwards
continue to be interections.  Additionally, new tests ensure that a ray which
is co-planar to one of the bounding box faces is not an interesection.